backend/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
config==0.4.2
cycler==0.10.0
Django==2.2.1
djangorestframework==3.9.4
docopt==0.6.2
idna==2.8
isort==4.3.18
kiwisolver==1.1.0
lazy-object-proxy==1.4.0
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
pipreqs==0.4.9
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.2
wincertstore==0.2
wrapt==1.11.1
yarg==0.1.9